+++ src/renderer/components/about-us/AboutUsPage.tsx
@@ -12,15 +12,15 @@ export const AboutUsPage = () => {
 
     return (
         <Box p={2}>
-            <Text fontSize='14px' color='textColor'>
+            <Text fontSize='14px'>
                 This app was build because our team was searching for a tool to see the number of our current open merge requests and make them easily
                 accessible. We didn't found such a tool and so we decided to build one by our own ;-)
             </Text>
 
-            <Text mt={3} fontWeight='bold' color='textColor'>
+            <Text mt={3} fontWeight='bold'>
                 Thanks
             </Text>
-            <Text fontSize='14px' color='textColor'>
+            <Text fontSize='14px'>
                 We want to thank our employers (
                 <Link href='#' onClick={openUrl('https://codecentric.de')} color='linkColor'>
                     codecentric
@@ -32,10 +32,10 @@ export const AboutUsPage = () => {
                 ) to make it possible to develop this app in our +1 time.
             </Text>
 
-            <Text mt={3} fontWeight='bold' color='textColor'>
+            <Text mt={3} fontWeight='bold'>
                 Issues / Feedback
             </Text>
-            <Text fontSize='14px' color='textColor'>
+            <Text fontSize='14px'>
                 You found an issue or missed some feature? We are very keen about your feedback and appreciate any help. Please{' '}
                 <Link href='#' onClick={openUrl('https://github.com/codecentric/merge-request-notifier/issues/new')} color='linkColor'>
                     create an issue
@@ -43,7 +43,7 @@ export const AboutUsPage = () => {
                 on GitHub.
             </Text>
 
-            <Text mt={3} textAlign='right' color='textColor'>
+            <Text mt={3} textAlign='right'>
                 <small>Version: {appVersion}</small>
             </Text>
         </Box>
