+++ src/renderer/components/form/CheckboxInput.tsx
@@ -1,21 +1,22 @@
 import * as React from 'react'
 import { Box, Text } from 'rebass'
-import { Label, Input } from '@rebass/forms'
+import { Label, Checkbox } from '@rebass/forms'
 
-interface FormInputProps {
+interface FormCheckboxProps {
     id: string
     label: string
+    name: string
     info?: string | JSX.Element
     error?: string
     [htmlAttribute: string]: any
 }
 
-export const FormInput: React.FunctionComponent<FormInputProps> = ({ id, label, error, info, ...props }) => (
+export const CheckboxInput: React.FunctionComponent<FormCheckboxProps> = ({ id, label, error, info, ...props }) => (
     <Box mb={3}>
-        <Label fontWeight='bold' fontSize={1} htmlFor={id} mb={1} color='textColor'>
-            {label}
+        <Label fontWeight='bold' fontSize={1} htmlFor={id}>
+            <Checkbox id={id} name={name} {...props} />
+            <Text p='3.5px'>{label}</Text>
         </Label>
-        <Input fontSize={2} id={id} {...props} sx={{ background: 'lightBackground', color: 'textColor' }} />
         {(!!error || !!info) && (
             <Text mt={1} mb={2} fontSize={1} lineHeight={1.4} color={!!error ? 'red' : 'textColor'}>
                 {error || info}
