+++ src/renderer/components/form/FormInput.tsx
@@ -1,24 +1,21 @@
 import * as React from 'react'
 import { Box, Text } from 'rebass'
-import { Label, Checkbox } from '@rebass/forms'
+import { Label, Input } from '@rebass/forms'
 
-interface FormCheckboxProps {
+interface FormInputProps {
     id: string
     label: string
-    name: string
     info?: string | JSX.Element
     error?: string
     [htmlAttribute: string]: any
 }
 
-export const CheckboxInput: React.FunctionComponent<FormCheckboxProps> = ({ id, label, error, info, ...props }) => (
+export const FormInput: React.FunctionComponent<FormInputProps> = ({ id, label, error, info, ...props }) => (
     <Box mb={3}>
-        <Label fontWeight='bold' fontSize={1} htmlFor={id}>
-            <Checkbox id={id} name={name} {...props} />
-            <Text p='3.5px' color='textColor'>
-                {label}
-            </Text>
+        <Label fontWeight='bold' fontSize={1} htmlFor={id} mb={1}>
+            {label}
         </Label>
+        <Input fontSize={2} id={id} {...props} sx={{ background: 'lightBackground' }} />
         {(!!error || !!info) && (
             <Text mt={1} mb={2} fontSize={1} lineHeight={1.4} color={!!error ? 'red' : 'textColor'}>
                 {error || info}
