+++ src/renderer/components/layout/Main.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { Box } from 'rebass'
 
 export const Main: React.FunctionComponent = ({ children }) => (
-    <Box sx={{ overflowY: 'scroll' }} bg='defaultBackground'>
+    <Box sx={{ overflowY: 'scroll' }} bg='defaultBackground' color='textColor'>
         {children}
     </Box>
 )
