+++ src/renderer/components/settings/SettingsPage.tsx
@@ -4,22 +4,17 @@ import { Tab, TabList, TabPanel, TabPanels, Tabs } from '@reach/tabs'
 import { ConnectionSettings } from './ConnectionSettings'
 import { GeneralSettings } from './GeneralSettings'
 import { Text } from 'rebass'
-import { lightTheme } from '../theme-light'
-import { darkTheme } from '../theme-dark'
-import { useConfig } from '../../hooks/config'
 
 export const SettingsPage: React.FunctionComponent = () => {
-    const { config } = useConfig()
-    const borderColor = config.generalConfig.darkMode ? darkTheme.colors.textColor : lightTheme.colors.textColor
     return (
         <>
             <Tabs>
-                <TabList style={{ border: borderColor }}>
+                <TabList>
                     <Tab>
-                        <Text color='textColor'>Connection</Text>
+                        <Text>Connection</Text>
                     </Tab>
                     <Tab>
-                        <Text color='textColor'>General</Text>
+                        <Text>General</Text>
                     </Tab>
                 </TabList>
                 <TabPanels>
