+++ src/renderer/components/themes/theme-dark.ts
@@ -1,0 +1,24 @@
+import { defaultTheme } from './theme-default'
+
+export const darkTheme = {
+    ...defaultTheme,
+    colors: {
+        defaultBackground: '#555555',
+        siteBackground: '#555555',
+        lightBackground: '#676767',
+        lightgray: '#848484',
+        gray: '#B3B3B3',
+        shadow: '#adb0b3',
+
+        textColor: '#e8e9e9',
+        textColorTinted: '#c1c5c8',
+        linkColor: '#4ea7f5',
+
+        blue: '#3e93de',
+        primary: '#3e93de',
+        primaryDarkened: '#3e93de',
+        green: '#1fad56',
+        red: '#e67373',
+        lightred: '#806767',
+    },
+}
